requirements.txt
# Recipe AI System Dependencies

# Web Crawling
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI & Embeddings
openai>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.0

# Utilities
python-dotenv>=1.0.0

